# pizzaproject/src/main/resources/openapi/implements/1804_2.yaml
openapi: 3.0.0
info:
  title: Pizza project API
  version: 1.1.0
  description: |
    Pizza API allows you to remotely manage the pizzeria.
    ### Description of the project and requirements:
    * Authorization is required to work with the API. Authorization is done by the bearer-token 😎
servers:
  - url: http://localhost:8090/
paths:
...
  /api/cafe/add:
    post:
      summary: Create new Cafe
      description: create new cafe
      tags:
        - Cafe
      operationId: create cafe
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 'Pizza house'
                adress:
                  type: string
                  example: 'Magdeburg, Lange str. 20'
                email:
                  type: string
                  example: 'pizza_info@pizza.net'
                phone:
                  type: string
                  example: '+49-151-111-222-33'
                open_at:
                  type: string
                  example: '8:00'
                close_at:
                  type: string
                  example: '20:00'
                image:
                  type: string
                  example: 'c_cafe1.jpg'
              required:
                - title
                - adress
                - email
                - phone
                - open_at
                - close_at
      responses:
        '200' :
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cafe'
        '401':
          description: Access denied
        '404':
          description: Request failed - No items
...
  /api/cafe/menu/add/{cafeId}/{pizzaId}:
    post:
      description: Add pizza to cafes menu
      summary: Add pizza to menu
      tags:
        - Cafe
      operationId: Add to menu
      parameters:
        - $ref: '#/components/parameters/cafeId'
        - $ref: '#/components/parameters/pizzaId'
      responses:
        200 :
          description: OK
        404:
          description: Bad request parametr
        401:
          description: Access denied
...
  /api/ingredients/{id}:
    get:
      summary: Get ingredient
      description: Get one by id
      tags:
        - Ingredients
      operationId: get ingredient
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '401':
          description: Access denied
        '404':
          description: Request failed - No items
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      description: record id
      schema:
        type: integer
    pizzaId:
      in: path
      name: pizzaId
      required: true
      description: pizza id
      schema:
        type: integer
    cafeId:
      in: path
      name: cafeId
      required: true
      description: cafe id
      schema:
        type: integer
  schemas:
...
    TypeOfInredient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Meat'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          example: 32
        name:
          type: string
          example: 'Bacon'
        price:
          type: number
          example: '4.50'
        image:
          type: string
          example: 'i_bacon.jpg'
        type:
          $ref: '#/components/schemas/TypeOfInredient'
...
    Cafe:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Pizza house'
        adress:
          type: string
          example: 'Magdeburg, Lange str. 20'
        email:
          type: string
          example: 'pizza_info@pizza.net'
        phone:
          type: string
          example: '+49-151-111-222-33'
        open_at:
          type: string
          example: '8:00'
        close_at:
          type: string
          example: '20:00'
        image:
          type: string
          example: 'c_cafe1.jpg'
security:
  - bearerAuth: []